refactor: split viewmodel in files viewmodel and codeattribute viewmodel
