+++ src/jvmMain/kotlin/ui/Controls.kt
@@ -10,30 +10,36 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import viewmodel.DebuggerViewModel
+import viewmodel.CodeAttributeViewmodel
 import viewmodel.Display
 
 @Composable
-fun Controls(viewModel: DebuggerViewModel, setShowFilePicker: (Boolean) -> Unit) {
-    Row(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.wrapContentSize(align = Alignment.Companion.CenterStart, unbounded = true)) {
+fun Controls(viewModel: CodeAttributeViewmodel?, setShowFilePicker: (Boolean) -> Unit) {
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        modifier = Modifier.wrapContentSize(align = Alignment.Companion.CenterStart, unbounded = true),
+    ) {
         Button(onClick = { setShowFilePicker(true) }) {
             Text("Open file")
         }
 
-        OutlinedButton(enabled = viewModel.hasPrevious, onClick = { viewModel.previous() }) {
+        OutlinedButton(enabled = viewModel != null && viewModel.hasPrevious, onClick = { viewModel?.previous() }) {
             Text("Previous")
         }
 
-        OutlinedButton(enabled = viewModel.hasNext, onClick = { viewModel.next() }) {
+        OutlinedButton(enabled = viewModel != null && viewModel.hasNext, onClick = {
+            viewModel?.next()
+        }) {
             Text("Next")
         }
 
-        OutlinedButton(enabled = viewModel.codeAttribute != null, onClick = {
-            viewModel.switchDisplay()
+        OutlinedButton(enabled = viewModel != null, onClick = {
+            viewModel?.switchDisplay()
         }) {
-            when (viewModel.display) {
+            when (viewModel?.display) {
                 Display.EVALUATIONS -> Text("Show results")
                 Display.RESULTS -> Text("Show evaluations")
+                else -> { Text("Show results") }
             }
         }
     }
