+++ src/jvmMain/kotlin/ui/stateview/StateViewer.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import viewmodel.DebuggerViewModel
+import viewmodel.CodeAttributeViewmodel
 import viewmodel.Display
 
 /**
@@ -15,41 +15,55 @@ import viewmodel.Display
  * Showing the current instruction, the stack, the variables and the branches that still need to be evaluated.
  */
 @Composable
-fun StateViewer(viewModel: DebuggerViewModel) {
+fun StateViewer(viewModel: CodeAttributeViewmodel?) {
     Column(Modifier.fillMaxSize()) {
-        when (viewModel.display) {
-            Display.RESULTS -> viewModel.instruction?.let { Instruction(it) }
-            Display.EVALUATIONS -> viewModel.evaluation?.let { InstructionEvaluation(it) }
+        viewModel?.let { viewModel ->
+            when (viewModel.display) {
+                Display.RESULTS -> viewModel.instruction?.let { Instruction(it) }
+                Display.EVALUATIONS -> viewModel.evaluation?.let { InstructionEvaluation(it) }
+            }
         }
 
         Category("Variables", maxHeight = 0.3F) {
-            when (viewModel.display) {
-                Display.RESULTS -> DisplayList(viewModel.instruction?.finalVariablesBefore ?: emptyList())
-                Display.EVALUATIONS -> DisplayList(viewModel.evaluation?.variablesBefore ?: emptyList())
+            if (viewModel != null) {
+                when (viewModel.display) {
+                    Display.RESULTS -> DisplayList(viewModel.instruction?.finalVariablesBefore ?: emptyList())
+                    Display.EVALUATIONS -> DisplayList(viewModel.evaluation?.variablesBefore ?: emptyList())
+                }
+            } else {
+                DisplayList<String>(emptyList())
             }
         }
 
         Row(horizontalArrangement = Arrangement.spacedBy(16.dp)) {
             Category("Stack", maxWidth = 0.7F) {
-                when (viewModel.display) {
-                    Display.EVALUATIONS -> DisplayList(viewModel.evaluation?.stackBefore?.reversed() ?: emptyList())
-                    Display.RESULTS -> DisplayList(viewModel.instruction?.finalStackBefore ?: emptyList())
+                if (viewModel != null) {
+                    when (viewModel.display) {
+                        Display.EVALUATIONS -> DisplayList(viewModel.evaluation?.stackBefore?.reversed() ?: emptyList())
+                        Display.RESULTS -> DisplayList(viewModel.instruction?.finalStackBefore ?: emptyList())
+                    }
+                } else {
+                    DisplayList<String>(emptyList())
                 }
             }
             Category("Branches") {
-                when (viewModel.display) {
-                    Display.EVALUATIONS -> DisplayList(
-                        viewModel.evaluationBlock?.branchEvaluationStack?.map { it.startOffset } ?: emptyList(),
-                    )
+                if (viewModel != null) {
+                    when (viewModel.display) {
+                        Display.EVALUATIONS -> DisplayList(
+                            viewModel.evaluationBlock?.branchEvaluationStack?.map { it.startOffset } ?: emptyList(),
+                        )
 
-                    Display.RESULTS -> {
-                        viewModel.instruction?.let { instruction ->
-                            DisplayList(
-                                (instruction.finalTargetInstructions ?: emptyList()) +
-                                    (instruction.finalOriginInstructions ?: emptyList()),
-                            )
+                        Display.RESULTS -> {
+                            viewModel.instruction?.let { instruction ->
+                                DisplayList(
+                                    (instruction.finalTargetInstructions ?: emptyList()) +
+                                        (instruction.finalOriginInstructions ?: emptyList()),
+                                )
+                            }
                         }
                     }
+                } else {
+                    DisplayList<String>(emptyList())
                 }
             }
         }
